Refactor OptimizeIndexing to make it more extensible
This PR decouples the logic necessary to compute bounds on variables
from the logic that uses this info to perform the strenght analysis on
int64 variables. While doing so, it tries to minimize the number of
attributes of the class in favour of local variables.

This class is now accessible from any `LoopBody` object.

ghstack-source-id: 245366a1edc4eced2a8b04a676d0eab1ca5aa0fb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100549