Fix bounds for sympy expressions
The analysis for SymPy expressions was incorrect as, even though it said
that the assumption was "smoothness" the assumtion was, in fact, that he
formula was monotone in every variable. In other words, it was
assuming that the derivative does not change signs in any variable (!!).

These issues showed up when using this analysis on a larger set of
formulas, where we would get bounds where lower > upper when using
ModularIndexing.

This PR extends the bounds analysis to SymPy expressions. It also
separates the symbols occuring on SymPy expressions into their own
class, as we don't want to use any fallback in this case.

ghstack-source-id: cb4af1659dd2bab36e8bd805d5cd1844eb22ff2a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102722